requirements.txt
dash
plotly
pandas
numpy
scikit-learn
supabase
python-dotenv
gunicorn
